refactor: move TypePriorities into TypeStrategy namespace
Move Flow\ETL\Schema\SortingStrategy\TypePriorities to Flow\ETL\Schema\SortingStrategy\TypeStrategy\TypePriorities per review feedback on #2459.
