+++ src/core/etl/src/Flow/ETL/DSL/functions.php
@@ -212,8 +212,8 @@ use Flow\ETL\Schema\SortingStrategy;
 use Flow\ETL\Schema\SortingStrategy\AlphabeticalStrategy;
 use Flow\ETL\Schema\SortingStrategy\CombinedStrategy;
 use Flow\ETL\Schema\SortingStrategy\MetadataStrategy;
-use Flow\ETL\Schema\SortingStrategy\TypePriorities as SchemaTypePriorities;
 use Flow\ETL\Schema\SortingStrategy\TypeStrategy;
+use Flow\ETL\Schema\SortingStrategy\TypeStrategy\TypePriorities as SchemaTypePriorities;
 use Flow\ETL\Schema\Validator\EvolvingValidator;
 use Flow\ETL\Schema\Validator\SelectiveValidator;
 use Flow\ETL\Schema\Validator\StrictValidator;
